requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
black==22.6.0
cfgv==3.3.1
click==8.1.3
decorator==5.1.1
distlib==0.3.5
Django==4.0.6
djangorestframework==3.13.1
executing==0.8.3
filelock==3.7.1
identify==2.5.2
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nodeenv==1.7.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
virtualenv==20.15.1
wcwidth==0.2.5
